Added: Authorization

+++ functions/src/handlers/parts.ts
@@ -4,6 +4,7 @@ import { logger } from "firebase-functions";
 import e = require("express");
 
 interface Part {
+  user: string;
   projectId: string;
   name: string;
   previous?: string;
@@ -20,24 +21,42 @@ const createPart = async (
   res: express.Response
 ): Promise<void> => {
   try {
-    if (!req.body.projectId || !req.body.name) {
+    if (!req.body.name) {
       res.status(400).json({ error: "Empty field detected" });
     } else {
-      const newPart: Part = {
-        projectId: req.params.projectId,
-        name: req.body.name,
-        previous: req.body.previous,
-        next: req.body.next,
-        description: req.body.description,
-        row: req.body.row,
-        currentRow: req.body.currentRow,
-        currentRowStitches: req.body.currentRowStitches,
-        createdAt: new Date().toISOString(),
-      };
-      const addedPart = await db.collection("parts").add(newPart);
+      const project = await db
+        .collection(`/projects`)
+        .doc(req.params.projectId)
+        .get();
+      if (project.exists) {
+        const resProject = {
+          projectId: project.id,
+          ...project.data(),
+        };
+        if ((<any>resProject).user === (<any>req).user.email) {
+          const newPart: Part = {
+            user: (<any>req).user.email,
+            projectId: req.params.projectId,
+            name: req.body.name,
+            previous: req.body.previous,
+            next: req.body.next,
+            description: req.body.description,
+            row: req.body.row,
+            currentRow: req.body.currentRow,
+            currentRowStitches: req.body.currentRowStitches,
+            createdAt: new Date().toISOString(),
+          };
 
-      const resPart = { ...newPart, partId: addedPart.id };
-      res.json(resPart);
+          const addedPart = await db.collection("parts").add(newPart);
+
+          const resPart = { ...newPart, partId: addedPart.id };
+          res.json(resPart);
+        } else {
+          res.status(404).json({ error: "Project's not found" });
+        }
+      } else {
+        res.status(404).json({ error: "Project's not found" });
+      }
     }
   } catch (error) {
     logger.error(error);
@@ -50,19 +69,35 @@ const getAllParts = async (
   res: express.Response
 ): Promise<void> => {
   try {
-    logger.log("Get All triggered");
-    const parts = await db
-      .collection("parts")
-      .where("projectId", "==", req.params.projectId)
+    // logger.log("Get All triggered");
+    const project = await db
+      .collection(`/projects`)
+      .doc(req.params.projectId)
       .get();
-    const partsList: Object[] = [];
-    parts.forEach((part) => {
-      partsList.push({
-        partId: part.id,
-        ...part.data(),
-      });
-    });
-    res.json(partsList);
+    if (project.exists) {
+      const resProject = {
+        projectId: project.id,
+        ...project.data(),
+      };
+      if ((<any>resProject).user === (<any>req).user.email) {
+        const parts = await db
+          .collection("parts")
+          .where("projectId", "==", req.params.projectId)
+          .get();
+        const partsList: Object[] = [];
+        parts.forEach((part) => {
+          partsList.push({
+            partId: part.id,
+            ...part.data(),
+          });
+        });
+        res.json(partsList);
+      } else {
+        res.status(404).json({ error: "Project's not found" });
+      }
+    } else {
+      res.status(404).json({ error: "Project's not found" });
+    }
   } catch (error) {
     logger.error(error);
     res.status(500).json({ error: error.code });
@@ -74,14 +109,18 @@ const getPart = async (
   res: express.Response
 ): Promise<void> => {
   try {
-    logger.log("Hello there", req.params);
     const part = await db.collection(`/parts`).doc(req.params.partId).get();
-    if (part.data()) {
+    if (part.exists) {
       const resPart = {
         partId: part.id,
         ...part.data(),
       };
-      res.json(resPart);
+      // logger.log("Hello there", resPart);
+      if ((<any>resPart).user === (<any>req).user.email) {
+        res.json(resPart);
+      } else {
+        res.status(404).json({ error: "Part's not found" });
+      }
     } else {
       res.status(404).json({ error: "Part's not found" });
     }
@@ -97,16 +136,23 @@ const updatePart = async (
 ): Promise<void> => {
   try {
     const part = await db.collection(`/parts`).doc(req.params.partId).get();
-    if (part.data()) {
+    if (part.exists) {
+      const oldPart = {
+        ...part.data(),
+      };
       const newPart = {
         ...req.body,
         updatedAt: new Date().toISOString(),
       };
-      await db
-        .collection("/parts")
-        .doc(req.params.partId)
-        .set(newPart, { merge: true });
-      res.json({ message: "Part's updated successfully" });
+      if ((<any>oldPart).user === (<any>req).user.email) {
+        await db
+          .collection("/parts")
+          .doc(req.params.partId)
+          .set(newPart, { merge: true });
+        res.json({ message: "Part's updated successfully" });
+      } else {
+        res.status(404).json({ error: "Part's not found" });
+      }
     } else {
       res.status(404).json({ error: "Part's not found" });
     }
@@ -121,8 +167,20 @@ const deletePart = async (
   res: express.Response
 ): Promise<void> => {
   try {
-    await db.collection("parts").doc(req.params.partId).delete();
-    res.status(204).send("Part's deleted successfully");
+    const part = await db.collection(`/parts`).doc(req.params.partId).get();
+    if (part.exists) {
+      const oldPart = {
+        ...part.data(),
+      };
+      if ((<any>oldPart).user === (<any>req).user.email) {
+        await db.collection("parts").doc(req.params.partId).delete();
+        res.status(204).send("Part's deleted successfully");
+      } else {
+        res.status(404).json({ error: "Part's not found" });
+      }
+    } else {
+      res.status(404).json({ error: "Part's not found" });
+    }
   } catch (error) {
     logger.error(error);
     res.status(500).json({ error: error.code });
@@ -135,15 +193,19 @@ const changeCurrentRow = async (
 ): Promise<void> => {
   try {
     const part = await db.collection("/parts").doc(req.params.partId).get();
-    if (part.data()) {
+    if (part.exists) {
       const updatedPart = { ...part.data() };
       updatedPart.currentRow++;
       updatedPart.updatedAt = new Date().toISOString();
-      await db
-        .collection("/parts")
-        .doc(req.params.partId)
-        .set(updatedPart, { merge: true });
-      res.json({ message: "Row's added successfully" });
+      if ((<any>updatedPart).user === (<any>req).user.email) {
+        await db
+          .collection("/parts")
+          .doc(req.params.partId)
+          .set(updatedPart, { merge: true });
+        res.json({ message: "Row's added successfully" });
+      } else {
+        res.status(404).json({ error: "Part's not found" });
+      }
     } else {
       res.status(404).json({ error: "Part's not found" });
     }
